moved some steps into tags and split ssh steps into a new file
